Now playing UI changes: Initial commit
This is a work in progress, aimed to rework the Now Playing screen with a couple of changes:

The album art is now full-bleed to the top of the screen. A text protection scrim is used to make the contents of the toolbar visible.

The ‘up next’ queue is no longer shown via a FAB, but rather an additional sliding panel at the bottom of the screen.

Much of the work for this fairly major UI change has been undertaken in this commit, however, the landscape mode is completely nerfed. I’ve got some designs/ideas for how the now playing screen should look in landscape mode, but it’s quite different from portrait, so there’s a fair bit of work to be done. The unfinished landscape stuff is a pretty big hurdle for finalising this change.

There are still some bugs to iron out as well, including:

1.) The toolbar doesn’t always respond to the panel slide, sometimes it remains opaque when it should be transparent

2.) The queue seems to dissappear, not exactly sure under what circumstances at the moment.