Adjustments to search precision..
When performing a Jaro-Winkler string search, the fields being searched are now split at whitespace. The score of the best-matching substring is then returned.

This means that when searching for “door” in “the wolf at the door”, we get a score of 1.0 instead of, say, 0.5 (or whatever it was before), since ‘door’ is a 100% match for one of the substrings in the title.