Fixed an issue when generating album-artists from albums..
When building the list of albums from songs, we need to ensure that songs with null album tags are not grouped with other songs with null album tags, if those other songs don’t have the same album-artist.

See the comments at Operators.songsToAlbums().

Also added Unit Tests for the Operators class, to ensure it behaves as expected.

Resolves #52.