Don’t show genre song count if there are >150 genres
Resolved #79.

A cursor is created for each genre entry in order to determine the number of songs. Unfortunately, this is the only way to determine the song count for a genre, because of the way the MediaStore is written.

For users who have tons of genres (usually due to messy id3 tags and the MediaStore not cleaning up empty genres), this can cause slow loading of the library, as there is a limit to the number of cursors which can be open at any given time.