Changes to genre behavior subject..
Previously, a cursor was opened for every genre retrieved in the getGenresRelay() call. This cursor is used to retrieve the number of songs in a genre.

Creating this many cursors in such a short timeframe can have a fairly negative impact on the performance of all queries, on some devices.

The only place in the app where the genre song count is truly required, is in ShuttleApplication.cleanGenres(). The song count is used to determine whether a genre can be deleted. Instead of running all genre song count cursors at once when cleaning genres, these are now spawned on a 250ms delay.

The genre song counts have been removed from the GenreFragment for now.

Fixes #79