Fixed an issue where enqueued songs were not added to the not-in-use-playlist
Whene MusicService.enqueue is called, the songs were only being added to the current playlist (if shuffle is on, the shuffle playlist, else the regular playlist). So, when the user enqueues songs, then changes shuffle mode, the enqueued songs were lost.

Also necessary to listen for the QUEUE_CHANGED event in the player presenter, to update the current queue position & queue length.