Fixed an issue where service wasn’t re-bound after being killed by system
The new background service excution limits on Android will kill a service pretty quickly after the app is considered ‘idle’, if the Service is not in the foreground.

When Shuttle is launched, Activities bind to the MusicService in onCreate(). Once Shuttle is minimised, if no music is playing, so the service is not in the foreground, the service will be killed. Upon reopening the Activity, the service is no longer alive (or bound), despite the service token still being valid.

This fix just ensures the service token is set to null if the service is disconnected, and the service is re-bound when the Activity is resumed, if required.

Resolves https://github.com/timusus/Shuttle/issues/136