Keep library tab controller’s views in memory
Store the rootView / recyclerView for the main tabs (AlbumArtistFragment, AlbumFragment, etc) as a member instance variable, and don’t recreate them in onCreateView() if they’re non-null.

This results in higher memory usage, but means we’re not recreating all these controllers when navigating back from the detail screens. This is a worthwhile trade-off, as the main library screens are very expensive to create.

Solves an issue where frames are dropped when navigating back from detail controller.

Resolves #180