Various work on album-shuffle mode
Album shuffle has been dialled back a bit - it’s no longer a fully-fledged shuffle mode on the now playing screen, but rather just an option at the top of the album listing screen, and an artist overflow menu item.

We might promote this feature in other places (like the artist detail screen) in the future.

This reverts a bunch of changes in the implementation of album-shuffle mode, and simplifies things. See Operators.albumShuffleSongs() for the actual songs to album-shuffled songs implementation.