Fixed an issue where play/pause is out of sync in notifications
The notification was calling MusicUtils.isPlaying() which returns false if the MusicService is not bound to an activity. Instead, pass in the current isPlaying() state into the notification helper (which is only ever called from the service anyway).

Resolves #151

Also fixed an issue where pressing play didn’t start the service on Oreo, if Shuttle isn’t already running in the background. This is resolved by using PendingIntent.getForegroundService instead of PendingIntent.getService (on Oreo).