Possible crash fix when refreshing SuggestedFragmemt items
Add both the retrieval of items disposable, and the set items disposable to the CompositeDisposable, and clear it when ‘refreshAdapter()’ is called.

This should prevent refreshAdapter() async callbacks from returning while another call is in-flight, which results in a sort of race condition leading to to possibly the latter call returning first, resulting in invalid adapter item positions.