Work on In IAB..
Added a PurchasePresenter/PurchaseView, which can be eubclassed by and Presenter/View wishing to display the upgrade to Shuttle+ dialog.

Moved IAB handling to the BaseActivity. Call IabManager.getInstance() when MainActivity launches, to setup the IAB manager.

Show the upgrade dialog when choosing the folder browser from the free version of the app. (Previously, the drawer item was disabled and nothing happened on click)