Finer-grained management of playlist menu observables
This commit breaks the makePlaylistMenu() into two methods. createPlaylistMenu() and createUpdatingPlaylistMenu().

The updating playlist menu method returns a Completable which can be disposed of wherever is suitable.

This is preferavle to the LifeCycleProvider in certain cases where the Completable needs to be disposed of outside of Android lifecycle events. For example, in some cases the menus are recreated when the ‘user visible hint’ changes. We need to dispose of the existing playlist menu observable before creating a new one. Passing in a LifeCycleProvider Observable does not allow us to handle this case.

Where the playlist menu is not required to ‘auto update’, we now just call createPlaylistMenu(), which only takes the first emission when querying the playlists relay - meaning we don’t need to manually dispose the Observable.

While I really like the idea of the LifecycleProvider emitting lifecycle events, I think this commit better clarifies the intention of the code.