Fixed a crash when pressing ChromeCast button
Attempting to turn off ChromeCast shows a dialog (MediaRouteControllerDialog). This dialog was throwing an InflateException on inflation, due to a missing image src attribute on an ImageButton.

It looks like this was somehow related to the parent of the ImageButton being created via reflection - and possibly attrs being lost somewhere along the way. This reflection stuff was only useful so that Aesthetic could apply automatic-theming to the view, something which has been removed in Shuttle anyway (since it causes memory leaks).

Resolves #126