Fixed an issue with PlayerFragment UI state when resuming from landscape
When the app is backgrounded in portrait, and the phone is forced into lanscape mode, if the track changes, the background view in the PlayerFragment is updated with blurred artwork. When returning to the app, the views are not updated again (because onResume is called, without an onCreateView). 

This means that the blurred artwork ends up displaying in portrait mode, resulting in what looks like a UI glitch.

Resolves #141