Fixed an issue where favorite icon was out of sync
Remove useless 'invalidatePlaylistRelay()' code, instead use Observable to monitor for changes to favorites - and convert to single if only the latest value is required.

Added a delay to the subscription which calls the success consumer when addint to / removing from favourites - so we're not saying the action was successful until we've given the underlying database/behavior relay time to update itself - so they're queried on success, their values will be current.