Added a launch task to fix year value for songs in MediaStore
Resolves #289

When the app is launched, songs with a zero value for year are iterated over. The tags are retrieved, and if the year is non-zeo, the entry is updated in the MediaStore.

This also resolves a bug where similar tasks happening on launch were executed immediately, instead of on a delay, due to an RxJava issue (need to defer the completable task, otherwise it is executed as soon as subscribe is called)