Possible crash fix for PlayerFragment
Ensure we still have a context, and we're added before attempting to apply color changes which require a context.

There is a cleaner way to do this, but we are cancelling the animation in onDestroyView(), yet it's still running in some situations where the context is null. Perhaps there's a timing issue to do with storing only one animation object despite running multiple animations - so maybe we're cancelling only one of them.