Potential fix for issue where playback randomly pauses..
At the end of onStartCommand(), we call scheduleDelayedShutdown(). This is meant to be followed by calls to cancel the shutdown. When launching the app and playing music, these calls to cancel the shutdown come after the call to schedule the shutdown, which works just fine. However, when launching the service from a media button command, widget, or bluetooth, playback is started, and cancels the scheduled shutdown, _then_ the shutdown is scheduled.

This is resolved by making a check in scheduleShutdown() to ensure that music is not currently playing, the service isn't in use, and we haven't been paused due to an incoming call (transient loss of focus).

- Added some more logging