Fixed a crash after the last track ends..
Release the media player and set it to null in setNextDataSource(). Ensures that when the last track completes, we call trackEnded() with 'trackDidChange=false' (playback is complete, rather than having moved on to the next track). This in turn prevents a call to notify meta-data change, which then caused the playback position to be queried, resulting in an error as the media player is no longer initialised.