Added a dummy notification helper..
When the MusicService is started, the a dummy notification saying "Shuttle service running..." is shown, and the app is moved into the foreground state. This ensures that the system does not cause an ANR if we don't otherwise call through to stopService() (for example when we've decided not to foreground the service, because it's about to be shut down).

The dummy notification is silent, and is removed 12.5 seconds after it is created (unless removed earlier). According to the source on github, the Android system will ANR your app if you don't call startForeground 10 seconds after starting the foreground service. So, we wait an extra 2.5 seconds before we shut down.

If Shuttle is then moved into the foreground because it actually needs to be, the dummy notification is removed, as it's no longer required.

When the MusicService is destroyed, the dummy notification is removed. It seems that it's OK for this to happen within the ANR delay, as long as we have called startForeground() at some point.

Thia seems like the least annoying way to resolve this issue.